+++ bundle/src/test/java/com/adobe/acs/commons/email/impl/EmailServiceImplTest.java
@@ -36,7 +36,7 @@ import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import javax.activation.DataSource;
 import javax.mail.internet.MimeMultipart;