+++ bundle/src/test/java/com/adobe/acs/commons/httpcache/store/jcr/impl/writer/BucketNodeFactoryTest.java
@@ -34,7 +34,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import com.adobe.acs.commons.httpcache.keys.CacheKey;
 import com.adobe.acs.commons.httpcache.store.jcr.impl.exceptions.BucketNodeFactoryException;
@@ -45,12 +45,6 @@ public class BucketNodeFactoryTest
 
     @Mock Session session;
     @Mock Node cacheRootNode;
-
-    @Before
-    public void setup() throws Exception{
-        when(session.isLive()).thenReturn(true);
-        when(cacheRootNode.getPath()).thenReturn(MockSettings.VALID_ROOT_PATH);
-    }
 
     @Test(expected = BucketNodeFactoryException.class)
     public void testInvalidPath() throws BucketNodeFactoryException, RepositoryException