+++ bundle/src/test/java/com/adobe/acs/commons/mcp/impl/processes/DataImporterTest.java
@@ -46,8 +46,8 @@ import org.mockito.invocation.InvocationOnMock;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.mockito.Matchers.any;
-import static org.mockito.Matchers.anyInt;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.spy;
@@ -110,7 +110,7 @@ public class DataImporterTest {
     }
 
     private Object runImmediately(InvocationOnMock invocation) {
-        Runnable r = invocation.getArgumentAt(0, Runnable.class);
+        Runnable r = invocation.getArgument(0);
         r.run();
         return null;
     }