+++ bundle/src/test/java/com/adobe/acs/commons/workflow/synthetic/impl/DeprecatedSyntheticGraniteWorkflowRunnerImplTest.java
@@ -37,13 +37,13 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import javax.jcr.Session;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -164,7 +164,6 @@ public class DeprecatedSyntheticGraniteWorkflowRunnerImplTest {
 
     @Test
     public void testExecute_Terminate() throws Exception {
-        when(session.hasPendingChanges()).thenReturn(true).thenReturn(false);
 
         Map<Object, Object> map = new HashMap<Object, Object>();
 
@@ -217,7 +216,6 @@ public class DeprecatedSyntheticGraniteWorkflowRunnerImplTest {
 
     @Test
     public void testExecute_Complete_noSave() throws Exception {
-        when(session.hasPendingChanges()).thenReturn(true).thenReturn(false);
 
         Map<Object, Object> map = new HashMap<Object, Object>();