+++ CHANGELOG.md
@@ -20,10 +20,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - #2090 - A failing on-deploy script could still have some of its pending changes persisted by the framework.
 - #2078 - Using the WorkflowPackageManager required read access to /(var/etc)/workflow/packages (fixes #2019)
 - #2120 - Fixed issues on the JCRHttpCacheStore regarding expiration handling, rewrote testcases (addresses #2113)
+- #2104 - Updated test suite to use mockito 3, build now working with Java 11
 
 ### Changed
 =======
 - #2101 - Cleanup public API of the remote Assets feature (#2094)
+
+
+
 
 ## [4.3.4] - 2019-10-16